Detects changes into repositories divided by init commit
We need to obtain the current state of the repository from a set of old references. To do this, we will use a list of changes to do per each first commit. These first commits represents our root repository, the way we will organize all the source code for all the repositories. The Changes could be create, update or delete.