Fix "reference not found error"
Repositories with no 'master' branch was failing because, trying to get the last commit date,  we had been trying to get a master reference that does not exists.
Now the logic is a little bit different: We iterate over all hash references, getting the time of each Commit that are pointing to. We return the newest one date.
References that are pointing to Tags are resolved. References pointing to Objects that are not Commits or Tags are ignored, printing a warning on logs.