dep prune
Added dep prune for test files and unused packages.

There is a more aggressive pruning option for non-go files, but I think that can be tricky with licenses, since it would remove the LICENSE/COPYING/README files of vendored projects, which might actually be a license violation for some projects.

Signed-off-by: Santiago M. Mola <santi@mola.io>