borges: check if siva file is empty when it shouldn't
If there are mentions to the init commit in any reference it checks that
the siva is not empty (contains at least one commit).

To do this two new functions are added to RepositoryStore interface:

* GetRefsByInit: Gets all references that have the provided init commit.
Still not used but will be needed soon to fix problems with missing or
broken siva files.
* InitHasRefs: Returns true if there's a reference with an specific init
commit. Similar to GetRefsByInit but faster because it only check that
at least one reference has it and does not retrieve repository data.

Note: we need an index in Reference.Init

Signed-off-by: Javi Fontan <jfontan@gmail.com>