borges: add fast path to single rooted repositories
When a repository only has one root and is the first one to write to it
the push step can be skipped and the packfile sent as is. This makes the
process quite faster at the expense of disk space.

* Rename references to add the repository ID
* Change configuration (adds repository remote)
* Delete remote "origin" used to clone the repository
* Create siva file and copy temporary repository to it
* Use copier to send the new siva file to the remote location

This process skips RootedTransactioner to download and upload siva
files. To be able to use Copier it is added to Archiver.

Also moved logging of time spent to copy siva files from
pushChangesToRootedRepository to tidy it a bit.

Signed-off-by: Javi Fontan <jfontan@gmail.com>